Changed the file strucutre to accommodate more platform specific files
