+++ Engine/include/Trundle/Platform/Linux/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(linux_include_files
+  window.h
+)
+
+target_sources(engine PRIVATE ${linux_include_files})
