+++ Engine/include/Trundle/Platform/MacOS/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(macos_include_files
+  window.h
+)
+
+target_sources(engine PRIVATE ${macos_include_files})
