+++ Engine/include/Trundle/Platform/Windows/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(windows_include_files
+  window.h
+)
+
+target_sources(engine PRIVATE ${windows_include_files})
