+++ Engine/src/Platform/Linux/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(linux_source_files
+  window.cpp
+)
+
+target_sources(engine PRIVATE ${linux_source_files})
