+++ Engine/src/Platform/MacOS/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(macos_source_files
+  window.cpp
+)
+
+target_sources(engine PRIVATE ${macos_source_files})
