+++ Engine/src/Platform/Windows/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(windows_source_files
+  window.cpp
+)
+
+target_sources(engine PRIVATE ${windows_source_files})
