+++ Engine/src/Platform/Windows/window.cpp
@@ -1,4 +1,4 @@
-//===-- linuxWindow.cpp ---------------------------------------------------===//
+//===-- window.cpp --------------------------------------------------------===//
 //
 // Copyright 2020 Zachary Selk
 //
@@ -15,14 +15,14 @@
 // limitations under the License.
 //
 //===----------------------------------------------------------------------===//
-#if defined(TRUNDLE_OS_LINUX)
+#if defined(TRUNDLE_OS_WINDOWS) // No longer needed
 
 #include <Trundle/Core/log.h>
 #include <Trundle/Events/keyEvent.h>
 #include <Trundle/Events/mouseEvent.h>
 #include <Trundle/Events/windowEvent.h>
 #include <Trundle/Platform/OpenGL/openGLContext.h>
-#include <Trundle/Platform/linuxWindow.h>
+#include <Trundle/Platform/Windows/window.h>
 
 namespace Trundle {
 
@@ -35,16 +35,16 @@ static void GLFWErrorCallback(int error, const char* description) {
 }
 
 Window* Window::create(const WindowProperties& properties) {
-  return new LinuxWindow(properties);
+  return new WindowsWindow(properties);
 }
 
-LinuxWindow::LinuxWindow(const WindowProperties& properties) {
+WindowsWindow::WindowsWindow(const WindowProperties& properties) {
   init(properties);
 }
 
-LinuxWindow::~LinuxWindow() { shutdown(); }
+WindowsWindow::~WindowsWindow() { shutdown(); }
 
-void LinuxWindow::init(const WindowProperties& properties) {
+void WindowsWindow::init(const WindowProperties& properties) {
   data.title = properties.title;
   data.width = properties.width;
   data.height = properties.height;
@@ -165,14 +165,14 @@ void LinuxWindow::init(const WindowProperties& properties) {
   });
 }
 
-void LinuxWindow::shutdown() { glfwDestroyWindow(window); }
+void WindowsWindow::shutdown() { glfwDestroyWindow(window); }
 
-void LinuxWindow::onUpdate() {
+void WindowsWindow::onUpdate() {
   glfwPollEvents();
   context->swapBuffers();
 }
 
-void LinuxWindow::setVSync(bool enable) {
+void WindowsWindow::setVSync(bool enable) {
   if (enable) {
     glfwSwapInterval(1);
   } else {
@@ -181,9 +181,9 @@ void LinuxWindow::setVSync(bool enable) {
   data.vSync = enable;
 }
 
-bool LinuxWindow::isVSync() const { return data.vSync; }
+bool WindowsWindow::isVSync() const { return data.vSync; }
 
-void* LinuxWindow::getNativeWindow() const { return (void*)window; }
+void* WindowsWindow::getNativeWindow() const { return (void*)window; }
 
 } // namespace Trundle
